Safari-611.1.10/LayoutTests/media/controls/showControlsButton-expected.txt

This tests for the existence of a hidden Show Controls button.

This test only runs in DRT!



Test for Show Controls button

EVENT: canplaythrough
PASS: We are using the apple idiom

Get the button that shows controls while in Voice Over

PASS: Name is 'Show Controls'
PASS: Is hidden

Testing finished.
